requirements.txt
playsound==1.2.2
pydub==0.25.1
python-dotenv==0.21.1
requests==2.28.2
websocket_client==1.5.1
websockets==10.4
websocket-client~=1.5.1
librosa~=0.10.0.post2
keyboard~=0.13.5
